Entity Id 속성 변경

See merge request s09-final/S09P31B306!14

+++ backend/homework/src/main/java/com/eefy/homework/domain/homework/persistence/entity/HomeworkSubmit.java
@@ -20,14 +20,14 @@ public class HomeworkSubmit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "homework_submit_id")
-    private Long id;
+    private Integer id;
 
     @OneToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "homework_question_id")
     private HomeworkQuestion homeworkQuestion;
 
     @Column(nullable = false)
-    private Long memberId;
+    private Integer memberId;
 
     @Column
     private Integer score;
